requirements.txt
itemadapter==0.6.0
pymongo==4.1.1
python-dotenv==0.20.0
Scrapy==2.6.1 
textblob==0.17.1
